requirements.txt
requests
python-dotenv
qrcode